- add a check for mysql driver to test 2 - add a test for postgres db
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@3198 689a6050-402a-0410-94f2-e92a70836424

+++ test/unit/2.sh
@@ -3,6 +3,11 @@
 
 # Needs a default openser database setup for mysql
 
+if [ ! -e ../modules/mysql/mysql.so ] ; then
+	echo "mysql driver not found, not run"
+	exit 0
+fi ;
+
 CFG=2.cfg
 cp $CFG $CFG.bak
